+++ Source/OxyPlot/Series/StemSeries.cs
@@ -119,8 +119,6 @@ namespace OxyPlot.Series
 
             double minDistSquared = this.MinimumSegmentLength * this.MinimumSegmentLength;
 
-            var clippingRect = this.GetClippingRect();
-
             // Transform all points to screen coordinates
             // Render the line when invalid points occur
             var dashArray = this.ActualDashArray;
@@ -139,8 +137,7 @@ namespace OxyPlot.Series
 
                 if (this.StrokeThickness > 0 && this.ActualLineStyle != LineStyle.None)
                 {
-                    rc.DrawClippedLine(
-                        clippingRect,
+                    rc.DrawReducedLine(
                         points,
                         minDistSquared,
                         actualColor,
@@ -159,7 +156,6 @@ namespace OxyPlot.Series
             if (this.MarkerType != MarkerType.None)
             {
                 rc.DrawMarkers(
-                    clippingRect,
                     markerPoints,
                     this.MarkerType,
                     this.MarkerOutline,